Add files via upload
Utiization of spaCy to perform advanced text preprocessing.  This included tasks of tokenization and lemmatization to transform raw text data into a manageable form. TF-IDF vectorization converted the preprocessed text into numerical features to represent the importance of words in the reviews relative to the dataset.  Each model was then trained and evaluated on the TF-IDF features which were split into training and testing sets.  Speed in processing through machine learning models varied greatly.  This is influenced by how effective the  model functions.  The accuracy across models on the test set varied.  Linear Regression and TF-IDF both achieved 0.88  LGBM Classifier was lower at .86.  BERT was unable to be run without a graphics card locally.  This project was performed without tuning hyperparameters to improve performance.  In the future if class imbalances were explored it would improve the robustness of models. The methods here serve as a foundation for further research into analyzing large data sets.