Add files via upload
Sweet Lift Taxi company has collected historical data on taxi orders at airports. To attract more drivers during peak hours, a model was built to predict the amount of taxi orders for the next hour. GridSearchCV was able to provide the best hyperparameters for the different types of models being utilized. Training set evaluated Random Forest, Decision Tree, Gradient Boosting, and LightGBM. In analyzing performance on the test set the best model was Gradient Boosting with an RMSE of 27.54. Grdient Boosting was then run with the test set data. The final RMSE was 27.54.